services/test_runner/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
redis>=5.2.0
pyyaml>=6.0.2
pydantic>=2.10.0
pytest>=8.3.0
httpx>=0.28.0
